[2020-08-30] Create Event API_09 : Polishing Source Code
